requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
psutil==5.9.6
python-socketio==5.8.0
eventlet==0.33.3
nvidia-ml-py==13.580.82
PyYAML==6.0.1
requests==2.31.0

